DraggableTile component breakout, event handler props
