+++ src/webapp/features/rabble/components/SwapTiles.tsx
@@ -91,6 +91,7 @@ const SwapTiles = ({
           onTileClick={sendTileToSwap}
           tilesInRack={tilesInRack}
           playerTiles={playerTiles}
+          onTileDrop={() => console.log("Swap Drop")}
         />
       </div>
     </div>
